// src/types/late-contributions-made.d.ts
export interface LateContributionsMade {
  name_of_contributor: string;
  city: string;
  state_zip: string;
  id_number: string;
  employer: string;
  occupation: string;
  amount: string;
  transaction_type: string;
  type: string;
  transaction_date: string;
  filed_date: string;
  transaction_number: string;
}